编写leetcode-152

+++ LeetCodeCpp/LeetCodeCpp/leetcode-215.cpp
@@ -9,7 +9,7 @@ public:
 		priority_queue<int, vector<int>, greater<int>> pq;
 		for (int i = 0; i < (int)nums.size(); ++i) {
 			pq.push(nums[i]);
-			if (pq.size() > k) pq.pop();
+			if ((int)pq.size() > k) pq.pop();
 		}
 		return pq.top();
 	}
